requirement.txt
streamlit==1.33.0
numpy==1.26.4
tensorflow==2.16.1
Pillow==10.3.0